backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
mysql-connector-python==8.2.0
sqlalchemy==2.0.23
pydantic[email]==2.5.0
python-dotenv==1.0.0